+++ wallet/test/integration/Test/Integration/Scenario/Transactions.hs
@@ -4,12 +4,95 @@ module Test.Integration.Scenario.Transactions
 
 import           Universum
 
+import qualified Data.List.NonEmpty as NonEmpty
+
 import qualified Cardano.Wallet.Client.Http as Client
 import           Test.Integration.Framework.DSL
-
+import           Test.Integration.Framework.Scenario (Scenario)
 
 spec :: Scenarios Context
 spec = do
+
+    -- estimated fee amount for this transaction is 187946
+    multioutputTransactionScenario
+        "proper fragmentation of utxo - both utxos can seperately cover fee and outputs"
+        [200000, 200000]
+        [10,11]
+        [ expectTxStatusEventually [InNewestBlocks] ]
+
+    -- estimated fee amount for this transaction is 187946
+    multioutputTransactionScenario
+        "proper fragmentation of utxo - 2 utxos available, each cannot seperately cover fee and outputs, but jointly can"
+        [100000, 100000]
+        [10,11]
+        [ expectTxStatusEventually [InNewestBlocks] ]
+
+    -- estimated fee amount for this transaction is 196076
+    multioutputTransactionScenario
+        "proper fragmentation of utxo - 3 utxos available, each cannot seperately cover fee and outputs, but jointly can"
+        [70000, 70000, 70000]
+        [10,11]
+        [ expectTxStatusEventually [InNewestBlocks] ]
+
+    -- estimated fee amount for this transaction is 187946
+    multioutputTransactionScenario
+        "proper fragmentation of utxo - 2 utxos available, one can cover fee, cannot cover any output seperately, but jointly can"
+        [187950, 50]
+        [10,11]
+        [ expectTxStatusEventually [InNewestBlocks] ]
+
+    multioutputTransactionScenario
+        "not enough fragmentation of utxo although available utxo can cover both outputs and fee"
+        [400000]
+        [10,11]
+        [ expectWalletError (UtxoNotEnoughFragmented (Client.ErrUtxoNotEnoughFragmented 1 Client.msgUtxoNotEnoughFragmented)) ]
+
+    multioutputTransactionScenario
+        "not enough fragmentation of utxo and available utxo cannot cover the sum of outputs and fee"
+        [100000]
+        [10,11]
+        [ expectWalletError (UtxoNotEnoughFragmented (Client.ErrUtxoNotEnoughFragmented 1 Client.msgUtxoNotEnoughFragmented)) ]
+
+
+    scenario "cannot send subsequent transaction when the first one is pending" $ do
+        fixtureSource <- setup $ defaultSetup
+            & initialCoins .~ [10000000]
+            & rawPassword .~ "raw password"
+
+        fixtureDest <- setup $ defaultSetup
+
+        -- Running two transactions one after another. Not waiting for the first transaction to be "completed".
+        -- The second transaction returns UtxoNotEnoughFragmented because the first one is still "pending"
+        resp1 <- request $ Client.postTransaction $- Payment
+            (defaultSource fixtureSource)
+            (defaultDistribution 1 fixtureDest)
+            defaultGroupingPolicy
+            (Just $ fixtureDest ^. spendingPassword)
+        verify resp1
+            [ expectSuccess
+            ]
+
+        resp2 <- request $ Client.postTransaction $- Payment
+            (defaultSource fixtureSource)
+            (defaultDistribution 1 fixtureDest)
+            defaultGroupingPolicy
+            (Just $ fixtureDest ^. spendingPassword)
+        verify resp2
+            [ expectWalletError (UtxoNotEnoughFragmented (Client.ErrUtxoNotEnoughFragmented 1 Client.msgUtxoNotEnoughFragmented))
+            ]
+
+        -- only after the first transaction completes the next one can be successfully sent
+        expectTxStatusEventually [InNewestBlocks, Persisted] resp1
+
+        resp3 <- request $ Client.postTransaction $- Payment
+            (defaultSource fixtureSource)
+            (defaultDistribution 1 fixtureDest)
+            defaultGroupingPolicy
+            (Just $ fixtureDest ^. spendingPassword)
+        verify resp3
+            [ expectTxStatusEventually [InNewestBlocks, Persisted]
+            ]
+
     scenario "successful payment appears in the history" $ do
         fixture <- setup $ defaultSetup
             & initialCoins .~ [1000000]
@@ -49,7 +132,7 @@ spec = do
             noSpendingPassword
 
         verify response
-            [ expectWalletError (NotEnoughMoney (ErrAvailableBalanceIsInsufficient 0))
+            [ expectWalletError (UtxoNotEnoughFragmented (Client.ErrUtxoNotEnoughFragmented 1 Client.msgUtxoNotEnoughFragmented))
             ]
 
     scenario "payment fails when wallet has insufficient funds" $ do
@@ -257,3 +340,39 @@ spec = do
             verify response
                 [ expectTxStatusEventually [InNewestBlocks, Persisted]
                 ]
+
+    multioutputTransactionScenario
+        :: String
+        -> [Word64]
+        -> [Word64]
+        -> [Either Client.ClientError Client.Transaction -> Scenario Context IO ()]
+        -> Scenarios Context
+    multioutputTransactionScenario title startCoins outputDistribution expectations =
+        scenario ("multi-output transaction: " <> title) $ do
+            fixtureSource <- setup $ defaultSetup
+                & initialCoins .~ startCoins
+
+            fixtureDest1 <- setup $ defaultSetup & walletName .~ "destinationWallet1"
+
+            accountDest1 <- successfulRequest $ Client.getAccount
+                $- (fixtureDest1 ^. wallet . walletId)
+                $- defaultAccountId
+
+            fixtureDest2 <- setup $ defaultSetup & walletName .~ "destinationWallet2"
+
+            accountDest2 <- successfulRequest $ Client.getAccount
+                $- (fixtureDest2 ^. wallet . walletId)
+                $- defaultAccountId
+
+            response <- request $ Client.postTransaction $- Payment
+                (defaultSource fixtureSource)
+                (customDistribution $
+                    NonEmpty.zipWith
+                    (,)
+                    (accountDest1 :| [accountDest2])
+                    (NonEmpty.fromList outputDistribution)
+                )
+                defaultGroupingPolicy
+                noSpendingPassword
+
+            verify response expectations